% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass.EMI.model.r
\name{carcass.EMI.simule}
\alias{carcass.EMI.simule}
\title{Wrapper function to the Carcass model for multiple sets of parameter values}
\usage{
carcass.EMI.simule(X, energy, PVi, duration)
}
\arguments{
\item{X}{: parameter matrix}

\item{energy}{: Metabolizable energy available}

\item{PVi}{: initial liveweight}

\item{duration}{: duration of simulation}
}
\value{
data.frame with PV,ProtC,ProtNC,LipC,LipNC
}
\description{
Wrapper function to the Carcass model for multiple sets of parameter values
}
