% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_data.r
\docType{data}
\name{carcass_data}
\alias{carcass_data}
\title{Data of growth of beef cattle for Carcass model}
\format{a \code{RangedData} instance, 1 row per plot.}
\source{
Agabriel, J. (com.pers.)
}
\usage{
carcass_data
}
\description{
This dataset is a list of 5 dataframe.
list_individuals : identifiant for each indivual
energy : Ration (type ration), Individu, time (week), energie (?)
init_condition : Individu, Pvi (initial liveweight)
observation_dynamic :Ration (3 levels "C","EM","F"), individu, time, PVobs
observation_slaughter : Ration, individu, time, PVVobs, PVobs, ProtCobs, ProtNCobs, LipCobs, LipNCobs.
}
