% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrot.weevil.model.r
\name{carrot.weevil.model}
\alias{carrot.weevil.model}
\title{Carrot weevil development model}
\usage{
carrot.weevil.model(tbase = 7, tteggs = 130, ttlarvae = 256,
  ttprepupae = 114, ttpupae = 130, ttadultpreovi = 91, weather,
  sdate = 1, ldate = 360)
}
\arguments{
\item{tbase}{: base temperature}

\item{tteggs}{: duration of eggs stage in degre.day}

\item{ttlarvae}{: duration of larvae stage in degre.day}

\item{ttprepupae}{: duration of prepupae stage in degre.day}

\item{ttpupae}{: duration of pupae stage in degre.day}

\item{ttadultpreovi}{: duration of adult stage until egg laying in degre.day}

\item{weather}{: weather data.frame for one single year}

\item{sdate}{: date to begin simulation (day of year) (default 1)}

\item{ldate}{: date to end simulation (day of year) (default 360)}
}
\value{
data.frame with daily state variable
}
\description{
\strong{Model description.} Simple model of developpement of carrot weevil.
}
