% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotton.model.r
\name{cotton.model}
\alias{cotton.model}
\title{The Cotton model (dynamic for numbers of Cotton fruiting points).}
\usage{
cotton.model(
  TESQ,
  PMAX,
  AFL,
  AL,
  AOP,
  P1,
  P2,
  P3,
  P4,
  P5,
  PF,
  PSF,
  TSQ,
  P,
  PR,
  PT,
  tend
)
}
\arguments{
\item{TESQ}{: time (in physiological days) at which square formation ends}

\item{PMAX}{: maximum number of squares}

\item{AFL}{: age of flower (physiological days)}

\item{AL}{: age of large boll (physiological days)}

\item{AOP}{: age of open boll (physiological days)}

\item{P1}{: rate of square formation before TF}

\item{P2}{: rate of square formation after TF}

\item{P3}{: shedding rate before TF}

\item{P4}{: shedding rate after TF}

\item{P5}{: boll shedding rate}

\item{PF}{: maximum number of squares for rate calculations}

\item{PSF}{: maximum number of setting bolls for rate calculations}

\item{TSQ}{: time (in physiological days) of first square}

\item{P}{: unused parameter}

\item{PR}{: unused parameter}

\item{PT}{: unused parameter}

\item{tend}{: end time of simulation (in physiological days)}
}
\value{
data.frame with daily state variable
}
\description{
\strong{Model description.} Dynamic model for numbers of Cotton fruiting points based on physiological time.
}
