\name{epirice.define.param}
\alias{epirice.define.param}
\title{Define values of the parameters for the Epirice model}
\usage{
  epirice.define.param()
}
\value{
  matrix with parameter values (nominal, binf, bsup)
}
\description{
  Define values of the parameters for the Epirice model
}

