% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epirice.model.r
\name{epirice.multi.simule}
\alias{epirice.multi.simule}
\title{Wrapper function to run Epirice multiple times (for multiple sets of inputs)}
\usage{
epirice.multi.simule(param, multi.simule, all = FALSE)
}
\arguments{
\item{param}{: a vector of parameters}

\item{multi.simule}{: matrix of n row definition of input variable : site, year and date of transplantation.}

\item{all}{: if you want a matrix combining multi.simule and output (default = FALSE)}
}
\value{
matrix with AUDPC for each input vector
}
\description{
Wrapper function for epirice.model
}
\seealso{
\code{\link{epirice.model}}
}
