% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epirice.model.r
\name{epirice.weather}
\alias{epirice.weather}
\title{Read weather data for Epirice (southern Asia weather)}
\usage{
epirice.weather(working.year = NA, working.site = NA,
  weather = weather_SouthAsia)
}
\arguments{
\item{working.year}{: year for the subset of weather data (default=NA : all the year)}

\item{working.site}{: site for the subset of weather data (default=NA : all the site)}

\item{weather}{: weather table}
}
\value{
data.frame with daily weather data for one or several site(s) and for one or several year(s)
}
\description{
Read weather data and format them for epirice.model
}
