\name{evaluation.criteria}
\alias{evaluation.criteria}
\title{Computation of classical evaluation criteria function}
\usage{
  evaluation.criteria(Ypred, Yobs, draw.plot = FALSE)
}
\arguments{
  \item{Ypred}{: prediction values from the model}

  \item{Yobs}{: observed values}

  \item{draw.plot}{: draw evaluation plot}
}
\value{
  data.frame with the different evaluation criteria
}
\description{
  Computation of classical evaluation criteria function
}

