\name{fRcW}
\alias{fRcW}
\title{Correction coefficient for infection - Relative humidity and rain (epirice)}
\usage{
  fRcW(RH, RAIN)
}
\arguments{
  \item{RH}{: relative humidity (percent)}

  \item{RAIN}{: precipitation (mm)}
}
\value{
  scalar for correction coefficient
}
\description{
  TODO
}

