% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness.of.fit.r
\name{goodness.of.fit}
\alias{goodness.of.fit}
\title{Calcule multiple goodness-of-fit criteria}
\usage{
goodness.of.fit(Yobs, Ypred, draw.plot = FALSE)
}
\arguments{
\item{Yobs}{: observed values}

\item{Ypred}{: prediction values from the model}

\item{draw.plot}{: draw evaluation plot}
}
\value{
data.frame with the different evaluation criteria
}
\description{
Calcule multiple goodness-of-fit criteria
}
\examples{
# observed and simulated values
obs<-c(78,110,92,75,110,108,113,155,150)
sim<-c(126,126,126,105,105,105,147,147,147)
goodness.of.fit(obs,sim,draw.plot=TRUE)
}
