\docType{data}
\name{grain.filling}
\alias{grain.filling}
\title{Wheat grain weight measurements after anthesis}
\format{a \code{RangedData} instance, 1 row per measurement. DD.aa = Degree Days after anthesis; grain.weight = Wheat grain weight (mg)}
\source{
  Darroch, B A, and R J Baker. 1990. "Grain filling in
  three spring wheat genotypes: statistical analysis." Crop
  Science 30(3):525-529.
}
\description{
  Darroch and Baker (1990) studied grain filling in three
  spring wheat genotypes. The data are seed weights of the
  spring wheat cultivar Neepawa in three different years
  1986, 1987, 1988. These data were numerised from figure
  of the article, so they present slight difference with
  original data.
}
\keyword{datasets}

