% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zadoks.original.model.r
\name{graph_epid}
\alias{graph_epid}
\title{Plot output of a Classical SEIR model for plant disease}
\usage{
graph_epid(out, typel = "s", all = TRUE, param = TRUE)
}
\arguments{
\item{out}{: output of the zadoks.original.model}

\item{typel}{: type of plot (default : s)}

\item{all}{: if all=true (default), plot all the state variable}

\item{param}{: if param (default), add the values of param on the plot}
}
\value{
plot
}
\description{
Plot the output of the Zadoks classical SEIR model for plant disease.
}
\seealso{
\code{\link{zakoks.original.model}}
}
