% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactation.calf.model.r
\name{lactation.calf.model2}
\alias{lactation.calf.model2}
\title{The Lactation model for use with lactation.calf.simule}
\usage{
lactation.calf.model2(param, duration, dt)
}
\arguments{
\item{param}{: a vector of parameters}

\item{duration}{: duration of simulation}

\item{dt}{: time step}
}
\value{
data.frame with CS, M, Mmoy, RM, day, week
}
\description{
see lactation.calf.model for model description.
}
\examples{
sim=lactation.calf.model2(lactation.define.param()["nominal",],6+2*7, 0.1)
}
