% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magarey.model.r
\name{magarey.model2}
\alias{magarey.model2}
\title{The Magarey model, taking a vector of parameters as argument}
\usage{
magarey.model2(T, param)
}
\arguments{
\item{T}{: input variable. Either a scalar or a vector (for a weather series).}

\item{param}{: parameters}
}
\value{
W : Wetness duration (hour). Either a scalar or a vector depending on T.
}
\description{
Generic model of infection for foliar diseases caused by fungi (from Magarey et al.,2005).
}
