% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.data_EuropeEU.r
\docType{data}
\name{maize.data_EuropeEU}
\alias{maize.data_EuropeEU}
\title{Maize Biomass and Leaf Area Data}
\format{
A \code{RangedData} instance, 1 row per measurement.
}
\source{
NA
}
\usage{
data(maize.data_EuropeEU)
}
\description{
Observation data for several sites and years in Europe.
This data are fake observed data, derived from simulations with an error model.
}
