% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.model.r
\name{maize.define.param}
\alias{maize.define.param}
\title{Define values of the parameters for the Maize model}
\usage{
maize.define.param()
}
\value{
matrix with parameter values (nominal, binf, bsup)
}
\description{
Define parameters values
}
