% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.model.r
\name{maize.model}
\alias{maize.model}
\title{The basic Maize model.}
\usage{
maize.model(Tbase, RUE, K, alpha, LAImax, TTM, TTL, weather, sdate, ldate)
}
\arguments{
\item{Tbase}{baseline temperature for growth (degreeC)}

\item{RUE}{radiation use efficiency (g.MJ-1)}

\item{K}{extinction coefficient (-)}

\item{alpha}{relative rate of LAI increase for small LAI ((degreeC.day)-1)}

\item{LAImax}{maximum leaf area index (m2 leaf/m2 soil)}

\item{TTM}{temperature sum for crop maturity (degreeC.day)}

\item{TTL}{temperature sum at end of leaf area increase (degreeC.day)}

\item{weather}{weather data.frame for one single year}

\item{sdate}{sowing date}

\item{ldate}{last date}
}
\value{
data.frame with daily TT, LAI, B
}
\description{
\strong{Model description.}
This model is a dynamic model of crop growth for Maize cultivated in
potential conditions. The crop growth is represented by three state
variables, leaf area per unit ground area (leaf area index, LAI), total
biomass (B) and cumulative thermal time since plant emergence (TT).
}
\details{
The tree state variables are dynamic variables depending on days
after emergence: TT(day), B(day), and LAI(day). The model has a time
step dt of one day.
\cr (1) \eqn{TT(day+1) = TT(day)+dTT(day)}
\cr (2) \eqn{B(day+1) = B(day)+dB(day)}
\cr (3) \eqn{LAI(day+1) = LAI(day)+dLAI(day)}
\cr (4) \eqn{dTT(day) = \max(\frac{TMIN+TMAX}{2}-Tbase;0)}
\cr (5) \eqn{dB(day) = RUE*(1-e^{-K*LAI})*I,\ if\ TT\le TTM;\  0\ otherwise}
\cr (6) \eqn{dLAI(day) = alpha*dTT*LAI*\max(LAImax-LAI;0),\ if\ TT\le TTL;\ 0\ otherwise}
}
\examples{
weather = maize.weather(working.year=2010, working.site=30,
  weather_all=weather_EuropeEU)
maize.model(Tbase=7, RUE=1.85, K=0.7, alpha=0.00243, LAImax=7,
  TTM=1200, TTL=700, weather, sdate=100, ldate=250)
}
\seealso{
\code{\link{maize.model2}}, \code{\link{maize.define.param}},
  \code{\link{maize.simule}}, \code{\link{maize.multisy}},
  \code{\link{maize.simule240}}, \code{\link{maize.simule_multisy240}}
}
