% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.muchow.model.r
\name{maize.muchow.graph}
\alias{maize.muchow.graph}
\title{Plot dynamic output of Muchow Maize model.}
\usage{
maize.muchow.graph(res)
}
\arguments{
\item{res}{: list of result from maize.muchow.model}
}
\description{
Plot 6 graphs of main output variables of the Muchow Maize model.
}
\examples{
# not run in package test
# res = maize.muchow.model(weather=maize.weather(working.year=2010, working.site=1))
# maize.muchow.graph(res)
}
\seealso{
\code{\link{mm.A.fct}}, \code{\link{mm.LN.fct}}, \code{\link{mm.FAS.fct}}, \code{\link{maize.multisy}},
\code{\link{mm.HI.fct}},\code{\link{maize.muchow.model}}
}
