% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.model.r
\name{maize.RUEtemp}
\alias{maize.RUEtemp}
\title{Calculate effect of temperature on RUE for Maize}
\usage{
maize.RUEtemp(T, RUE_max, T0, T1, T2, T3)
}
\arguments{
\item{T}{: temperature}

\item{RUE_max}{: maximum value for RUE}

\item{T0}{: temperature parameter}

\item{T1}{: temperature parameter}

\item{T2}{: temperature parameter}

\item{T3}{: temperature parameter}
}
\value{
RUE value
}
\description{
Function to compute effect of temperature on RUE
}
