% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.model.r
\name{maize_cir.model}
\alias{maize_cir.model}
\title{The Maize model with additional state variable CumInt}
\usage{
maize_cir.model(Tbase, RUE, K, alpha, LAImax, TTM, TTL, weather, sdate, ldate)
}
\arguments{
\item{Tbase}{baseline temperature for growth (degreeC)}

\item{RUE}{radiation use efficiency (g.MJ-1)}

\item{K}{extinction coefficient (-)}

\item{alpha}{relative rate of LAI increase for small LAI ((degreeC.day)-1)}

\item{LAImax}{maximum leaf area index (m2 leaf/m2 soil)}

\item{TTM}{temperature sum for crop maturity (degreeC.day)}

\item{TTL}{temperature sum at end of leaf area increase (degreeC.day)}

\item{weather}{weather data.frame for one single year}

\item{sdate}{sowing date}

\item{ldate}{last date}
}
\value{
data.frame with daily TT, LAI, B, CumInt
}
\description{
Variant of the maize model that also tracks cumulative
  intercepted radiation (CumInt). Internally delegates to the shared
  `.maize_engine()` with `track_cumint = TRUE`.
}
