\name{q.arg.fast.runif}
\alias{q.arg.fast.runif}
\title{Build the q.arg argument for the  FAST function (sensitivity analysis)}
\usage{
  q.arg.fast.runif(model.factors)
}
\arguments{
  \item{model.factors}{: matrix defining minimal (binf) and
  maximal values (bsup) for a set of p parameters}
}
\value{
  a list of list
}
\description{
  according to minimal and maximal values defined in a
  model.factors matrix
}

