% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedweight.data_data.r
\docType{data}
\name{seedweight.data}
\alias{seedweight.data}
\title{Wheat grain weight measurements after anthesis}
\format{a \code{RangedData} instance, 1 row per measurement. DD = Degree Days after anthesis; seedweight = Wheat grain weight (mg)}
\source{
Darroch, B A, and R J Baker. 1990. "Grain filling in three spring wheat genotypes: statistical analysis." Crop Science 30(3):525-529.
}
\usage{
seedweight.data
}
\description{
Darroch and Baker (1990) studied grain filling in three spring wheat genotypes.
The data are seed weights of the spring wheat cultivar Neepawa in three different years 1986, 1987, 1988.
These data were numerised from figure of the article, so they present slight difference with original data.
}
