% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness.of.fit.r
\name{threshold.measures}
\alias{threshold.measures}
\title{Computation of threshold.measures}
\usage{
threshold.measures(Yobs, Ypred, p, d, units = "")
}
\arguments{
\item{Yobs}{: observed values}

\item{Ypred}{: prediction values from the model}

\item{p}{: percentage threshold (0-100) for calculating TDIp}

\item{d}{: absolute error threshold value for calculating CP}

\item{units}{: units}
}
\value{
data.frame with the different evaluation criteria
}
\description{
Computation of threshold.measures
}
\examples{
# observed and simulated values
obs<-c(78,110,92,75,110,108,113,155,150)
sim<-c(126,126,126,105,105,105,147,147,147)
threshold.measures(obs,sim,80,1.0)
}
