% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verhulst.model.r
\name{verhulst.model}
\alias{verhulst.model}
\title{The Verhulst (logistic) model - calculate daily values over designated time period}
\usage{
verhulst.model(a, k, Y0, duration)
}
\arguments{
\item{a}{: growth rate}

\item{k}{: capacity}

\item{Y0}{: initial condition}

\item{duration}{: duration of simulation}
}
\value{
data.frame with daily Y
}
\description{
The Verhulst (logistic) model - calculate daily values over designated time period
}
\examples{
plot(verhulst.model(0.08,100,1,100), type="l", ylim=c(0,115),
   xlab="day", ylab="Y, population density",lwd=2)
}
\seealso{
\code{\link{verhulst.update}} for the update function of the Verhulst model.
}
