% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verhulst.model.r
\name{verhulst.update}
\alias{verhulst.update}
\title{The Verhulst (logistic) model - calculate change for one day}
\usage{
verhulst.update(Y, a, k)
}
\arguments{
\item{Y}{: state variable Y(t=day)}

\item{a}{: growth rate}

\item{k}{: capacity}
}
\value{
state variable at Y(t=day+1)
}
\description{
The Verhulst (logistic) model - calculate change for one day
}
\seealso{
\code{\link{verhulst.model}} for the integration loop function of the Verhulst model.
}
