\name{watbal.define.param}
\alias{watbal.define.param}
\title{Define values of the parameters for the WaterBalance model}
\usage{
  watbal.define.param()
}
\value{
  matrix with parameter values (nominal, binf, bsup)
}
\description{
  Define values of the parameters for the WaterBalance
  model
}

