% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watbal.model.r
\name{watbal.weather}
\alias{watbal.weather}
\title{Read weather data for the WaterBalance model (West of France Weather)}
\usage{
watbal.weather(working.year = NA, working.site = NA)
}
\arguments{
\item{working.year}{: year for the subset of weather data (default=NA : all the year)}

\item{working.site}{: site for the subset of weather data (default=NA : all the site)}
}
\value{
data.frame with daily weather data for one or several site(s) and for one or several year(s)
}
\description{
Read weather data for the WaterBalance model (West of France Weather)
}
