% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_GNS_data.r
\docType{data}
\name{weather_GNS}
\alias{weather_GNS}
\title{Weather series for Gainesville (FL, USA) years 1982 and 1983}
\format{a \code{RangedData} instance, 1 row per day.}
\source{
University of Florida
}
\usage{
weather_GNS
}
\description{
This  daily climate dataset for Gainesville (FL, USA) years 1982 and 1983
covers the period 1st January 1982 to 31 December 1983
with 2 complete years of data with precipitation.
The dataset was provide by JJW Jones and the university of Florida to run simulation of the
publication of Muchow et al. (1990). 
This climate dataset contains daily estimates of precipitation (RAIN), minimum (Tmin) and
maximum temperature (Tmax), solar radiation (I), 
photosynthically active radiation (PAR)
}
