% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weed.model.r
\name{weed.define.param}
\alias{weed.define.param}
\title{Define parameter values of the Weed model}
\usage{
weed.define.param()
}
\value{
matrix with parameter values (nominal, binf, bsup)
}
\description{
Define parameter values of the Weed model
}
