\name{weed.model}
\alias{weed.model}
\title{The weed model function}
\usage{
  weed.model(param, weed.deci)
}
\arguments{
  \item{param}{: vector of the 16 parameters}

  \item{weed.deci}{: decisions table for Soil, Crop et
  Herbicide}
}
\value{
  data.frame with annual values of yield
}
\description{
  TOCOMPLETE
}

