% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weed.model.r
\name{weed.simule}
\alias{weed.simule}
\title{Wrapper function to run the Weed model multiple times (for multiple sets of inputs)}
\usage{
weed.simule(X, weed.deci)
}
\arguments{
\item{X}{: parameter matrix}

\item{weed.deci}{: decisions table for Soil, Crop et Herbicide}
}
\value{
matrix with Yield for year 3 for each parameter vector
}
\description{
Wrapper function to run the Weed model multiple times (for multiple sets of inputs)
}
