\docType{data}
\name{Wheat_GPC}
\alias{Wheat_GPC}
\title{Grain Protein Contents in Wheat Grains}
\format{a \code{RangedData} instance, 1 row per plot.}
\source{
  Barbottin et al. 2008
}
\description{
  This dataset contains GPC values for 43 plots. The first
  column (GPC) corresponds to measured data and the two
  other corresponds to the results obtained by two models
  (GPC.model1 and GPC.model2).
}
\keyword{datasets}

