% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WheatYieldGreece_data.r
\docType{data}
\name{WheatYieldGreece}
\alias{WheatYieldGreece}
\title{National Wheat Yield evolution for Greece from FAO}
\format{a \code{RangedData} instance, 1 row per measurement. Year, Yield : Wheat Yield (hectogram/hectare = 0.0001 ton/hectare)}
\source{
Food And Agricultural Organization of United Nations (FAO), FAOSTAT database, http://faostat.fao.org
}
\usage{
WheatYieldGreece
}
\description{
Wheat yield time series data in Greece from 1961 to 2010 yield.
}
