% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotton.model.r
\name{cotton.model}
\alias{cotton.model}
\title{The Cotton model (dynamic for numbers of Cotton fruiting points).}
\usage{
cotton.model(TESQ, PMAX, AFL, AL, AOP, P1, P2, P3, P4, P5, PF, PSF, TSQ, P,
  PR, PT, tend)
}
\arguments{
\item{TESQ}{: TO COMPLETE}

\item{PMAX}{: TO COMPLETE}

\item{AFL}{: TO COMPLETE}

\item{AL}{: TO COMPLETE}

\item{AOP}{: TO COMPLETE}

\item{P1}{: TO COMPLETE}

\item{P2}{: TO COMPLETE}

\item{P3}{: TO COMPLETE}

\item{P4}{: TO COMPLETE}

\item{P5}{: TO COMPLETE}

\item{PF}{: TO COMPLETE}

\item{PSF}{: TO COMPLETE}

\item{TSQ}{: TO COMPLETE}

\item{P}{: TO COMPLETE}

\item{PR}{: TO COMPLETE}

\item{PT}{: TO COMPLETE}

\item{tend}{: TO COMPLETE}
}
\value{
data.frame with daily state variable
}
\description{
\strong{Model description.} TO COMPLETE
}
