% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.data_EuropeEU.r
\docType{data}
\name{maize.data_EuropeEU}
\alias{maize.data_EuropeEU}
\title{maize biomass and leaf area data}
\format{a \code{RangedData} instance, 1 row per measurement.}
\source{
NA
}
\usage{
maize.data_EuropeEU
}
\description{
"observation data" for several site and year (site-year) in Europe.
This data are fake observed date, derived from simulations with an error model.
}
