% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weed.model.r
\name{weed.model}
\alias{weed.model}
\title{The Weed model - calculate daily values over designated time period}
\usage{
weed.model(param, weed.deci)
}
\arguments{
\item{param}{: vector of the 16 parameters}

\item{weed.deci}{: decisions table for Soil, Crop et Herbicide}
}
\value{
data.frame with annual values of yield
}
\description{
The Weed model - calculate daily values over designated time period
}
