\name{weed.simule}
\alias{weed.simule}
\title{Wrapping function to run an virtual experimental design for weed model}
\usage{
  weed.simule(X, weed.deci)
}
\arguments{
  \item{X}{: parameter matrix}

  \item{weed.deci}{: decisions table for Soil, Crop et
  Herbicide}
}
\value{
  matrix with Yield for year 3 for each parameter vector
}
\description{
  Wrapping function to run an virtual experimental design
  for weed model
}

