% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wheat_GPC_data.r
\docType{data}
\name{Wheat_GPC}
\alias{Wheat_GPC}
\title{Grain Protein Contents in Wheat Grains}
\format{a \code{RangedData} instance, 1 row per plot.}
\source{
Barbottin et al. 2008
}
\usage{
Wheat_GPC
}
\description{
This dataset contains data for 43 plots. The column GPC corresponds to
measured data, GPC.model1 and GPC.model2 correspond to the results obtained by two models.
The other columns contain other information for each plot (technical practices) :
number, tillage,	max_water,	preceding_crop,	sow_date,	Nendwinter,	Nfertilizer,	SPAD,	NNI,	Yield.
}
